Refactor Answer.jsx to use CSS modules
