+++ client/src/Components/QuestionList/QuestionList.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'; // eslint-disable-line no-unused-vars
-import Question from './Question.jsx'; // eslint-disable-line no-unused-vars
+import Question from '../Question/Question.jsx'; // eslint-disable-line no-unused-vars
 
 const QuestionList = (props) => {
   const { questions } = props;
