+++ client/src/Components/style3.css
@@ -1,3 +1,3 @@
 .some {
-  background-color: red;
+  background-color: green;
 }
