+++ client/src/Components/styles.css
@@ -1,0 +1,101 @@
+button {
+  font-family: Arial;
+  border: 1px solid transparent;
+  font-weight: bold;
+  outline: none;
+  cursor: pointer;
+  border-radius: 2px;
+}
+
+.btn-primary {
+  color: #fff;
+  background-color: #00a680;
+  border-color: #00a680 #267060 #267060 #00a680;
+}
+
+.btn-primary:hover {
+  background-color: #00c094;
+}
+
+.btn-secondary {
+  color: #00a680;
+  background-color: #fff;
+  border-color: #e5e5e5 #767676 #767676 #e5e5e5;
+}
+
+.btn-secondary:hover {
+  border-color: #f2f2f2 #838383 #838383 #f2f2f2;
+  color: #00c094;
+}
+
+button.arrow {
+  display: block;
+  height: 20px;
+  width: 20px;
+  border-radius: 30px;
+  border: 2px solid #00a680;
+  color: #00a680;
+  text-align: center;
+  margin: auto;
+  font-weight: 1000;
+  position: relative;
+}
+
+button.arrow:hover {
+  color: #fff;
+  background-color: #00a680;
+}
+
+.fa-chevron-down {
+  top: 70%;
+}
+
+.fa-flag {
+  cursor: pointer;
+}
+
+button.big {
+  padding: 8px 16px;
+  font-size: 14px;
+  line-height: 18px;
+}
+
+button.small {
+  padding: 6px 10px;
+  font-size: 12px;
+  line-height: 16px;
+}
+
+a {
+  outline: none;
+  color: #069;
+  font-family: Arial;
+  text-decoration: none;
+}
+
+a:hover {
+  text-decoration: underline;
+}
+
+textarea {
+  border-radius: 5px;
+  border: 1px solid #dfdddd;
+  border-top-width: 3px;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 16px;
+  padding: 10px;
+  width: 730px;
+}
+
+textarea:focus {
+  outline: none !important;
+  border:1px solid #00a680;
+}
+
+body {
+  background-color: #f2f2f2;
+}
+
+.top-sum {
+  margin-bottom: -2px;
+}
