--- style.css
@@ -1,3 +1,0 @@
-.some {
-  background-color: red;
-}
