+++ webpack.config.js
@@ -1,6 +1,7 @@
 const path = require('path');
 // const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const combineLoaders = require('webpack-combine-loaders');
+// const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 // const StaticSiteGeneratorPlugin = require('static-site-generator-webpack-plugin');
 
 // const locals = {
